+++ datastore/inMemory/tenantServiceDatastoreInMemory.go
@@ -10,6 +10,7 @@ import (
 	uuid "github.com/satori/go.uuid"
 
 	"github.com/accedian/adh-gather/models/common"
+	metmod "github.com/accedian/adh-gather/models/metrics"
 	tenmod "github.com/accedian/adh-gather/models/tenant"
 )
 
@@ -17,12 +18,14 @@ import (
 // database operations for the Admin Service when using local memory
 // as the storage option. Useful for tests.
 type TenantServiceDatastoreInMemory struct {
-	tenantToIDtoTenantUserMap              map[string]map[string]*tenmod.User
-	tenantToIDtoTenantDomainMap            map[string]map[string]*tenmod.Domain
-	tenantToIDtoTenantConnectorConfigMap   map[string]map[string]*tenmod.ConnectorConfig
-	tenantToIDtoTenantConnectorInstanceMap map[string]map[string]*tenmod.ConnectorInstance
-	tenantToIDtoTenantMonitoredObjectMap   map[string]map[string]*tenmod.MonitoredObject
-	tenantToIDtoTenantThrPrfMap            map[string]map[string]*tenmod.ThresholdProfile
+	tenantToIDtoTenantUserMap                 map[string]map[string]*tenmod.User
+	tenantToIDtoTenantDomainMap               map[string]map[string]*tenmod.Domain
+	tenantToIDtoTenantConnectorConfigMap      map[string]map[string]*tenmod.ConnectorConfig
+	tenantToIDtoTenantConnectorInstanceMap    map[string]map[string]*tenmod.ConnectorInstance
+	tenantToIDtoTenantMonitoredObjectMap      map[string]map[string]*tenmod.MonitoredObject
+	tenantToIDtoTenantThrPrfMap               map[string]map[string]*tenmod.ThresholdProfile
+	tenantToIDtoTenantReportScheduleConfigMap map[string]map[string]*metmod.ReportScheduleConfig
+	tenantToIDtoTenantSLAReportMap            map[string]map[string]*metmod.SLAReport
 
 	tenantIDtoMetaSlice   map[string][]*tenmod.Metadata
 	tenantIDtoIngPrfSlice map[string][]*tenmod.IngestionProfile
@@ -38,6 +41,7 @@ func CreateTenantServiceDAO() (*TenantServiceDatastoreInMemory, error) {
 	res.tenantToIDtoTenantConnectorConfigMap = map[string]map[string]*tenmod.ConnectorConfig{}
 	res.tenantToIDtoTenantMonitoredObjectMap = map[string]map[string]*tenmod.MonitoredObject{}
 	res.tenantToIDtoTenantThrPrfMap = map[string]map[string]*tenmod.ThresholdProfile{}
+	res.tenantToIDtoTenantSLAReportMap = map[string]map[string]*metmod.SLAReport{}
 
 	res.tenantIDtoMetaSlice = map[string][]*tenmod.Metadata{}
 	res.tenantIDtoIngPrfSlice = map[string][]*tenmod.IngestionProfile{}
@@ -83,6 +87,14 @@ func (tsd *TenantServiceDatastoreInMemory) DoesTenantExist(tenantID string, ctx 
 		}
 	case tenmod.TenantThresholdProfileType:
 		if tsd.tenantToIDtoTenantThrPrfMap[tenantID] == nil {
+			return tenantDNE
+		}
+	case tenmod.TenantReportScheduleConfigType:
+		if tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID] == nil {
+			return tenantDNE
+		}
+	case tenmod.TenantSLAReportStr:
+		if tsd.tenantToIDtoTenantSLAReportMap[tenantID] == nil {
 			return tenantDNE
 		}
 	default:
@@ -1022,4 +1034,170 @@ func (tsd *TenantServiceDatastoreInMemory) BulkInsertMonitoredObjects(tenantID s
 	}
 
 	return result, nil
+}
+
+func (tsd *TenantServiceDatastoreInMemory) CreateReportScheduleConfig(config *metmod.ReportScheduleConfig) (*metmod.ReportScheduleConfig, error) {
+	if err := tsd.DoesTenantExist(config.TenantID, tenmod.TenantReportScheduleConfigType); err != nil {
+		tsd.tenantToIDtoTenantReportScheduleConfigMap[config.TenantID] = map[string]*metmod.ReportScheduleConfig{}
+	}
+
+	recCopy := metmod.ReportScheduleConfig{}
+	deepcopy.Copy(&recCopy, config)
+	recCopy.ID = uuid.NewV4().String()
+	recCopy.REV = uuid.NewV4().String()
+	recCopy.CreatedTimestamp = ds.MakeTimestamp()
+	recCopy.LastModifiedTimestamp = recCopy.CreatedTimestamp
+
+	tsd.tenantToIDtoTenantReportScheduleConfigMap[config.TenantID][recCopy.ID] = &recCopy
+
+	return &recCopy, nil
+}
+func (tsd *TenantServiceDatastoreInMemory) UpdateReportScheduleConfig(config *metmod.ReportScheduleConfig) (*metmod.ReportScheduleConfig, error) {
+	if len(config.ID) == 0 {
+		return nil, fmt.Errorf("%s must have an ID", tenmod.TenantReportScheduleConfigStr)
+	}
+	if len(config.REV) == 0 {
+		return nil, fmt.Errorf("%s must have a revision", tenmod.TenantReportScheduleConfigStr)
+	}
+	if err := tsd.DoesTenantExist(config.TenantID, tenmod.TenantReportScheduleConfigType); err != nil {
+		return nil, fmt.Errorf("%s does not exist", tenmod.TenantReportScheduleConfigStr)
+	}
+
+	recCopy := metmod.ReportScheduleConfig{}
+	deepcopy.Copy(&recCopy, config)
+	recCopy.REV = uuid.NewV4().String()
+	recCopy.LastModifiedTimestamp = ds.MakeTimestamp()
+
+	tsd.tenantToIDtoTenantReportScheduleConfigMap[config.TenantID][recCopy.ID] = &recCopy
+
+	return &recCopy, nil
+}
+func (tsd *TenantServiceDatastoreInMemory) DeleteReportScheduleConfig(tenantID string, configID string) (*metmod.ReportScheduleConfig, error) {
+	if len(configID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Report schedule config ID", tenmod.TenantReportScheduleConfigStr)
+	}
+	if len(tenantID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Tenant ID", tenmod.TenantReportScheduleConfigStr)
+	}
+	if err := tsd.DoesTenantExist(tenantID, tenmod.TenantReportScheduleConfigType); err != nil {
+		return nil, fmt.Errorf("%s does not exist", tenmod.TenantReportScheduleConfigStr)
+	}
+
+	rec, ok := tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID][configID]
+	logger.Log.Debugf(models.AsJSONString(tsd.tenantToIDtoTenantReportScheduleConfigMap))
+	if ok {
+		delete(tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID], configID)
+
+		if len(tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID]) == 0 {
+			delete(tsd.tenantToIDtoTenantReportScheduleConfigMap, tenantID)
+		}
+		return rec, nil
+	}
+
+	return nil, fmt.Errorf("%s not found", tenmod.TenantReportScheduleConfigStr)
+}
+func (tsd *TenantServiceDatastoreInMemory) GetReportScheduleConfig(tenantID string, configID string) (*metmod.ReportScheduleConfig, error) {
+	if len(configID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Report schedule config ID", tenmod.TenantReportScheduleConfigStr)
+	}
+	if len(tenantID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Tenant ID", tenmod.TenantReportScheduleConfigStr)
+	}
+	if err := tsd.DoesTenantExist(tenantID, tenmod.TenantReportScheduleConfigType); err != nil {
+		return nil, fmt.Errorf("%s does not exist", tenmod.TenantReportScheduleConfigStr)
+	}
+
+	rec, ok := tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID][configID]
+	if ok {
+		return rec, nil
+	}
+
+	return nil, fmt.Errorf("%s not found", tenmod.TenantReportScheduleConfigStr)
+}
+func (tsd *TenantServiceDatastoreInMemory) GetAllReportScheduleConfigs(tenantID string) ([]*metmod.ReportScheduleConfig, error) {
+	err := tsd.DoesTenantExist(tenantID, tenmod.TenantReportScheduleConfigType)
+	if err != nil {
+		return []*metmod.ReportScheduleConfig{}, nil
+	}
+
+	recList := make([]*metmod.ReportScheduleConfig, 0)
+
+	for _, rec := range tsd.tenantToIDtoTenantReportScheduleConfigMap[tenantID] {
+		recList = append(recList, rec)
+	}
+
+	return recList, nil
+}
+
+func (tsd *TenantServiceDatastoreInMemory) CreateSLAReport(slaReport *metmod.SLAReport) (*metmod.SLAReport, error) {
+	if err := tsd.DoesTenantExist(slaReport.TenantID, tenmod.TenantSLAReportType); err != nil {
+		tsd.tenantToIDtoTenantSLAReportMap[slaReport.TenantID] = map[string]*metmod.SLAReport{}
+	}
+
+	recCopy := metmod.SLAReport{}
+	deepcopy.Copy(&recCopy, slaReport)
+	recCopy.ID = uuid.NewV4().String()
+
+	tsd.tenantToIDtoTenantSLAReportMap[slaReport.TenantID][recCopy.ID] = &recCopy
+
+	return &recCopy, nil
+}
+
+func (tsd *TenantServiceDatastoreInMemory) DeleteSLAReport(tenantID string, slaReportID string) (*metmod.SLAReport, error) {
+	if len(slaReportID) == 0 {
+		return nil, fmt.Errorf("%s must provide a SLA Report ID", tenmod.TenantSLAReportStr)
+	}
+	if len(tenantID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Tenant ID", tenmod.TenantSLAReportStr)
+	}
+	if err := tsd.DoesTenantExist(tenantID, tenmod.TenantSLAReportType); err != nil {
+		return nil, fmt.Errorf("%s does not exist", tenmod.TenantSLAReportStr)
+	}
+
+	rec, ok := tsd.tenantToIDtoTenantSLAReportMap[tenantID][slaReportID]
+	logger.Log.Debugf(models.AsJSONString(tsd.tenantToIDtoTenantSLAReportMap))
+	if ok {
+		delete(tsd.tenantToIDtoTenantSLAReportMap[tenantID], slaReportID)
+
+		if len(tsd.tenantToIDtoTenantSLAReportMap[tenantID]) == 0 {
+			delete(tsd.tenantToIDtoTenantSLAReportMap, tenantID)
+		}
+		return rec, nil
+	}
+
+	return nil, fmt.Errorf("%s not found", tenmod.TenantSLAReportStr)
+}
+
+func (tsd *TenantServiceDatastoreInMemory) GetSLAReport(tenantID string, slaReportID string) (*metmod.SLAReport, error) {
+	if len(slaReportID) == 0 {
+		return nil, fmt.Errorf("%s must provide a SLA report ID", tenmod.TenantSLAReportStr)
+	}
+	if len(tenantID) == 0 {
+		return nil, fmt.Errorf("%s must provide a Tenant ID", tenmod.TenantSLAReportStr)
+	}
+	if err := tsd.DoesTenantExist(tenantID, tenmod.TenantSLAReportType); err != nil {
+		return nil, fmt.Errorf("%s does not exist", tenmod.TenantSLAReportStr)
+	}
+
+	rec, ok := tsd.tenantToIDtoTenantSLAReportMap[tenantID][slaReportID]
+	if ok {
+		return rec, nil
+	}
+
+	return nil, fmt.Errorf("%s not found", tenmod.TenantSLAReportStr)
+}
+
+func (tsd *TenantServiceDatastoreInMemory) GetAllSLAReports(tenantID string) ([]*metmod.SLAReport, error) {
+	err := tsd.DoesTenantExist(tenantID, tenmod.TenantSLAReportType)
+	if err != nil {
+		return []*metmod.SLAReport{}, nil
+	}
+
+	recList := make([]*metmod.SLAReport, 0)
+
+	for _, rec := range tsd.tenantToIDtoTenantSLAReportMap[tenantID] {
+		recList = append(recList, rec)
+	}
+
+	return recList, nil
 }