+++ datastore/tenantServiceDatastore.go
@@ -2,6 +2,7 @@ package datastore
 
 import (
 	"github.com/accedian/adh-gather/models/common"
+	metmod "github.com/accedian/adh-gather/models/metrics"
 	tenmod "github.com/accedian/adh-gather/models/tenant"
 )
 
@@ -59,4 +60,15 @@ type TenantServiceDatastore interface {
 	UpdateTenantMeta(meta *tenmod.Metadata) (*tenmod.Metadata, error)
 	DeleteTenantMeta(tenantID string) (*tenmod.Metadata, error)
 	GetTenantMeta(tenantID string) (*tenmod.Metadata, error)
+
+	CreateReportScheduleConfig(config *metmod.ReportScheduleConfig) (*metmod.ReportScheduleConfig, error)
+	UpdateReportScheduleConfig(config *metmod.ReportScheduleConfig) (*metmod.ReportScheduleConfig, error)
+	DeleteReportScheduleConfig(tenantID string, configID string) (*metmod.ReportScheduleConfig, error)
+	GetReportScheduleConfig(tenantID string, configID string) (*metmod.ReportScheduleConfig, error)
+	GetAllReportScheduleConfigs(tenantID string) ([]*metmod.ReportScheduleConfig, error)
+
+	CreateSLAReport(slaReport *metmod.SLAReport) (*metmod.SLAReport, error)
+	DeleteSLAReport(tenantID string, slaReportID string) (*metmod.SLAReport, error)
+	GetSLAReport(tenantID string, slaReportID string) (*metmod.SLAReport, error)
+	GetAllSLAReports(tenantID string) ([]*metmod.SLAReport, error)
 }