+++ monitoring/metrics.go
@@ -23,8 +23,10 @@ const (
 	ThrCrossStrTopN          = "thr_cross_topn"
 	HistogramStr             = "histogram"
 	RawMetricStr             = "raw_metric"
+	GenSLAReportStr          = "gen_sla_report"
 	SLAReportStr             = "sla_report"
 	TenantMetaStr            = "tenant_meta"
+	ReportSchedConfigStr     = "report_sched_conf"
 	AdminViewsStr            = "admin_views"
 	ValidTypesStr            = "valid_types"
 	AggMetricsStr            = "aggr_metrics"
@@ -123,14 +125,24 @@ const (
 	GetThrCrossByMonObjTopNStr = ThrCrossStr + metricNameDelimiter + MonitoredObjectStr + metricNameDelimiter + OPGetStr
 	GetHistogramObjStr         = HistogramStr + metricNameDelimiter + OPGetStr
 	GetRawMetricStr            = RawMetricStr + metricNameDelimiter + OPGetStr
-	GetSLAReportStr            = SLAReportStr + metricNameDelimiter + OPGetStr
-	QueryAggregatedMetricsStr  = AggMetricsStr + metricNameDelimiter + OPGetStr
+	GenerateSLAReportStr       = GenSLAReportStr + metricNameDelimiter + OPGetStr
+
+	QueryAggregatedMetricsStr = AggMetricsStr + metricNameDelimiter + OPGetStr
 
 	CreateTenantMetaStr = TenantMetaStr + metricNameDelimiter + OPCreateStr
 	UpdateTenantMetaStr = TenantMetaStr + metricNameDelimiter + OPUpdateStr
 	GetTenantMetaStr    = TenantMetaStr + metricNameDelimiter + OPGetStr
 	PatchTenantMetaStr  = TenantMetaStr + metricNameDelimiter + OPPatchStr
 	DeleteTenantMetaStr = TenantMetaStr + metricNameDelimiter + OPDeleteStr
+
+	CreateReportScheduleConfigStr = ReportSchedConfigStr + metricNameDelimiter + OPCreateStr
+	UpdateReportScheduleConfigStr = ReportSchedConfigStr + metricNameDelimiter + OPUpdateStr
+	GetReportScheduleConfigStr    = ReportSchedConfigStr + metricNameDelimiter + OPGetStr
+	GetAllReportScheduleConfigStr = ReportSchedConfigStr + metricNameDelimiter + OPGetAllStr
+	DeleteReportScheduleConfigStr = ReportSchedConfigStr + metricNameDelimiter + OPDeleteStr
+
+	GetSLAReportStr    = SLAReportStr + metricNameDelimiter + OPGetStr
+	GetAllSLAReportStr = SLAReportStr + metricNameDelimiter + OPGetAllStr
 
 	CreateValidTypesStr      = ValidTypesStr + metricNameDelimiter + OPCreateStr
 	UpdateValidTypesStr      = ValidTypesStr + metricNameDelimiter + OPUpdateStr